build: upgrade ESLint v8+typescript-eslint v6 to v9/v8 (#102)
* build: upgrade ESLint v8+typescript-eslint v6 to v9/v8

- Upgrade eslint 8.54.0 (EOL exact pin) to ^9.36.0
- Upgrade @typescript-eslint/* from ^6.12.0 to ^8.37.0
- Add @eslint/eslintrc ^3.3.1 for FlatCompat adapter
- Bump eslint-plugin-react 7.33.2 to ^7.37.0 (ESLint v9 peer support)
- Migrate .eslintrc.js to eslint.config.mjs (flat config)
- Remove .eslintignore (patterns in globalIgnores)
- Resolves GHSA-3ppc-4f35-3m26 GHSA-23c5-xmqv-rm74 GHSA-7r86-cg39-jmmj

Closes #96


* fix: address CodeRabbit review comments on eslint-v9 PR

- Use path.join() instead of string concatenation in getDirectories()
- Use string 'warn' instead of numeric 1 for import/order severity


* build: remove patch-package and postinstall-postinstall (yarn migration cleanup)

Neither package has any work to do after the yarn-to-npm migration:
- No patches/ directory exists, so patch-package is a no-op
- postinstall-postinstall was a yarn workaround; npm runs postinstall natively

The presence of postinstall-postinstall caused node-ci Install dependencies
to fail because it detected the global yarn binary on the runner and
attempted to run yarn, which was then rejected by Corepack (packageManager
is set to npm).


* build: add conventional-commits-filter@^5.0.0 to fix npm ci on npm 10

npm 10.9.8 (used by the GitHub Actions Node 22 runner) incorrectly
treats an optional peer dep version mismatch as a lockfile sync error.
@conventional-changelog/git-client@2.7.0 declares conventional-commits-filter@^5.0.0
as an optional peer dep, but only v4.0.0 was in the lockfile.

Adding v5 as a direct devDep ensures the lockfile satisfies npm 10's
validation while leaving v4 in place for packages that require it.


* fix: skip env validation during next build in CI

The shared node-ci workflow runs npm run build without setting the
required runtime env vars (NATS_SERVER_URL, ADMIN_SERVICE_URL,
TMS_SERVER_URL). Aligns with the existing pattern used by the test
script which already passes SKIP_ENV_VALIDATION=1.


* fix: exclude eslint.config.mjs from TypeScript compilation

tsconfig.json includes **/*.mjs, causing next build to type-check
eslint.config.mjs. The @typescript-eslint/parser package uses the
exports field for its type declarations, which is incompatible with
the project's moduleResolution: node setting.

eslint.config.mjs is a tooling config file and does not need to be
type-checked as part of the application build.


* fix: add SKIP_ENV_VALIDATION=1 to all next lint scripts

next lint loads next.config.mjs which imports env.mjs, triggering
env var validation. The CI runner does not have NATS_SERVER_URL,
ADMIN_SERVICE_URL, or TMS_SERVER_URL set, causing the lint step to fail.

Aligns all next lint scripts with the pattern already used by the
build and test scripts.


* fix: use dynamic import() for next-auth/jwt in server.js

next-auth v5 is ESM-only and cannot be loaded via require() in a
CommonJS module. Replace the top-level require with a lazy dynamic
import() inside the async Socket.IO auth middleware, where getToken
is already called in an async context.


* test: fix Playwright e2e test failures

- Fix stale page title in e2e/example.spec.ts (was 'Next.js Enterprise
  Boilerplate', actual title is 'Tazama Demo')
- Fix incorrect assertion in pipeline test: 'img[src*="neutral-light"].last()'
  can never pass because ProcessIndicator and DeviceInfo always keep neutral
  images visible; replaced with a scoped XPath locator that targets only the
  Event Adjudicator panel's StatusIndicator img
- Fix test comment: fixture result=500 < interdictionThreshold=600 so stop=false
  and color="y" (alert/yellow), not stop=true as previously stated
- Add NEXTAUTH_SECRET to playwright.config.ts webServer env to silence
  [auth][error] MissingSecret log spam in CI


---------
