+++ e2e/example.spec.ts
@@ -3,5 +3,5 @@ import { expect, test } from "@playwright/test"
 test("has title", async ({ page }) => {
   await page.goto("./")
 
-  await expect(page).toHaveTitle(/Next.js Enterprise Boilerplate/)
+  await expect(page).toHaveTitle(/Tazama Demo/)
 })
