+++ package.json
@@ -4,19 +4,19 @@
   "private": true,
   "scripts": {
     "dev": "cross-env FORCE_COLOR=1 nodemon server.js",
-    "build": "next build",
+    "build": "cross-env SKIP_ENV_VALIDATION=1 next build",
     "start": "node server.js",
     "clean": "npx rimraf .next node_modules coverage package-lock.json",
-    "lint": "next lint",
-    "lint:eslint": "next lint",
-    "lint:eslint-unused": "next lint --report-unused-disable-directives",
-    "lint:eslint-no-inline": "next lint --no-inline-config",
-    "lint:fix": "next lint --fix",
+    "lint": "cross-env SKIP_ENV_VALIDATION=1 next lint",
+    "lint:eslint": "cross-env SKIP_ENV_VALIDATION=1 next lint",
+    "lint:eslint-unused": "cross-env SKIP_ENV_VALIDATION=1 next lint --report-unused-disable-directives",
+    "lint:eslint-no-inline": "cross-env SKIP_ENV_VALIDATION=1 next lint --no-inline-config",
+    "lint:fix": "cross-env SKIP_ENV_VALIDATION=1 next lint --fix",
     "lint:prettier": "prettier --check \"**/*.{js,jsx,ts,tsx}\"",
     "prettier": "prettier --check \"**/*.{js,jsx,ts,tsx}\"",
     "prettier:fix": "prettier --write \"**/*.{js,jsx,ts,tsx}\"",
     "fix": "npm run fix:prettier && npm run fix:eslint",
-    "fix:eslint": "next lint --fix",
+    "fix:eslint": "cross-env SKIP_ENV_VALIDATION=1 next lint --fix",
     "fix:prettier": "prettier --write \"**/*.{js,jsx,ts,tsx}\"",
     "analyze": "cross-env ANALYZE=true npm run build",
     "prepare": "husky",
@@ -25,8 +25,7 @@
     "e2e:ui": "playwright test --ui",
     "format": "prettier --write \"**/*.{ts,tsx,md}\"",
     "preinstall": "npx npm-only-allow@1.2.6 --PM npm",
-    "postinstall": "npx patch-package -y",
-    "coupling-graph": "npx madge --extensions js,jsx,ts,tsx,css,md,mdx ./ --exclude '.next|tailwind.config.js|reset.d.ts|prettier.config.js|postcss.config.js|playwright.config.ts|next.config.js|next-env.d.ts|instrumentation.ts|e2e/|README.md|.storybook/|.eslintrc.js' --image graph.svg"
+    "coupling-graph": "npx madge --extensions js,jsx,ts,tsx,css,md,mdx ./ --exclude '.next|tailwind.config.js|reset.d.ts|prettier.config.js|postcss.config.js|playwright.config.ts|next.config.js|next-env.d.ts|instrumentation.ts|e2e/|README.md|.storybook/|eslint.config.mjs' --image graph.svg"
   },
   "dependencies": {
     "@next/bundle-analyzer": "^15.0.0",
@@ -78,16 +77,17 @@
     "@types/node": "^20.12.13",
     "@types/react": "^19.0.0",
     "@types/react-dom": "^19.0.0",
-    "@typescript-eslint/eslint-plugin": "^6.12.0",
-    "@typescript-eslint/parser": "^6.12.0",
+    "@eslint/eslintrc": "^3.3.1",
+    "@typescript-eslint/eslint-plugin": "^8.37.0",
+    "@typescript-eslint/parser": "^8.37.0",
     "all-contributors-cli": "^6.26.1",
     "autoprefixer": "^10.4.16",
     "cross-env": "^7.0.3",
-    "eslint": "8.54.0",
+    "eslint": "^9.36.0",
     "eslint-config-next": "15.5.18",
     "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-import": "^2.29.0",
-    "eslint-plugin-react": "7.33.2",
+    "eslint-plugin-react": "^7.37.0",
     "eslint-plugin-tailwindcss": "^3.13.0",
     "fetch-mock": "^9.11.0",
     "gzip-size": "6.0.0",
@@ -97,10 +97,9 @@
     "lint-staged": "^16.1.2",
     "minimatch": "^9.0.9",
     "mkdirp": "^3.0.1",
+    "conventional-commits-filter": "^5.0.0",
     "npm-only-allow": "^1.2.6",
-    "patch-package": "^8.0.0",
     "postcss": "^8.4.31",
-    "postinstall-postinstall": "^2.1.0",
     "prettier": "^3.6.2",
     "prettier-plugin-tailwindcss": "^0.5.7",
     "react-datepicker": "^7.3.0",
