+++ tsconfig.json
@@ -43,6 +43,7 @@
     ".next/types/**/*.ts"
   ],
   "exclude": [
-    "node_modules"
+    "node_modules",
+    "eslint.config.mjs"
   ]
 }
